PacBioKanga updates to improve throughput when utilising the distributed computing SW service capability. Added parameter to read error correcting submodule 'ecreads' allowing user to specify error correcting with sense overlap sense only. This allows user to independently assembly the sense/antisense strands of dsDNA.

+++ libbiokanga/SAMfile.cpp
@@ -2077,7 +2077,7 @@ int CSAMfile::CSIDepth(INT64 MaxSeqLen,			// max sequence length for any alignme
 {
 INT64 s;
 int n_lvls;
-for (n_lvls = 0, s = (INT64)(1 << min_shift); MaxSeqLen > s; ++n_lvls, s <<= 3);
+for (n_lvls = 0, s = (INT64)((INT64)1 << min_shift); MaxSeqLen > s; ++n_lvls, s <<= (INT64)3);
 return(n_lvls);
 }
